Revise📖 & Fix issue🔧 1. 📖Now images in `Gallery` are sorted by `changedTime` 2. 🔧Fixed issue that auto save button will lose statement #9
Former-commit-id: d16243f2c1abab258d4b44303e33554f03d1cc21

+++ lib/widgets/mercurius/mercurius_gallery_card_widget.dart
@@ -70,9 +70,7 @@ class MercuriusGalleryCardWidget extends StatelessWidget {
                             summary: l10n.pleaseThinkTwiceAboutDeletingTheImage,
                             context: context,
                           ).confirm;
-                          if (confirm == true) {
-                            fileSystemEntity.deleteSync();
-                          }
+                          if (confirm == true) fileSystemEntity.deleteSync();
                         },
                   icon: const Icon(
                     Icons.delete_outline_rounded,
